+++ src/js/core/core.js
@@ -1,24 +1,9 @@
-import { css, fastdom, inBrowser, observeResize, on, toMs } from 'uikit-util';
+import { css, inBrowser, on, toMs } from 'uikit-util';
 
-export default function (UIkit) {
+export default function () {
     if (!inBrowser) {
         return;
     }
-
-    // throttle 'resize'
-    let pendingResize;
-    const handleResize = () => {
-        if (pendingResize) {
-            return;
-        }
-        pendingResize = true;
-        fastdom.read(() => (pendingResize = false));
-        UIkit.update(null, 'resize');
-    };
-
-    on(window, 'load resize', handleResize);
-    on(document, 'loadedmetadata load', handleResize, true);
-    observeResize(document.documentElement, handleResize);
 
     let started = 0;
     on(