+++ src/js/core/leader.js
@@ -1,9 +1,10 @@
 import Class from '../mixin/class';
 import Media from '../mixin/media';
 import { attr, getCssVar, toggleClass, unwrap, wrapInner } from 'uikit-util';
+import Resize from '../mixin/resize';
 
 export default {
-    mixins: [Class, Media],
+    mixins: [Class, Media, Resize],
 
     props: {
         fill: String,
@@ -31,26 +32,19 @@ export default {
     },
 
     update: {
-        read({ changed, width }) {
-            const prev = width;
-
-            width = Math.floor(this.$el.offsetWidth / 2);
+        read() {
+            const width = Math.trunc(this.$el.offsetWidth / 2);
 
             return {
                 width,
                 fill: this.fill,
-                changed: changed || prev !== width,
                 hide: !this.matchMedia,
             };
         },
 
-        write(data) {
-            toggleClass(this.wrapper, this.clsHide, data.hide);
-
-            if (data.changed) {
-                data.changed = false;
-                attr(this.wrapper, this.attrFill, new Array(data.width).join(data.fill));
-            }
+        write({ width, fill, hide }) {
+            toggleClass(this.wrapper, this.clsHide, hide);
+            attr(this.wrapper, this.attrFill, new Array(width).join(fill));
         },
 
         events: ['resize'],