+++ src/js/core/index.js
@@ -3,7 +3,6 @@ export { default as Alert } from './alert';
 export { default as Cover } from './cover';
 export { default as Drop, default as Dropdown } from './drop';
 export { default as FormCustom } from './form-custom';
-export { default as Gif } from './gif';
 export { default as Grid } from './grid';
 export { default as HeightMatch } from './height-match';
 export { default as HeightViewport } from './height-viewport';